Adapted to rename, added executable scope
