+++ src/oracles/types.ts
@@ -1,5 +1,5 @@
 import { Goal, PpString } from '../lib/coq-lsp/types';
 
 export interface Oracle {
-  query(goal: Goal<PpString>): Promise<String>
+  query(goal: Goal<PpString>): Promise<string>
 }
